Removed compiler warnings & Default state
- Algorithm destructor was missing override keyword
- Algorithm constructor sets the default state for processors (Possibly error prone)
- Component mover in live constant wasn't stopping the timer when destroyed.
- Unused variables have explicitly been set to unused for juce look n feel
- Casting and minor type problems resolved

+++ atesh_module/PluginProcessor.cpp
@@ -64,7 +64,7 @@ void AteshAudioProcessor::changeProgramName (int , const juce::String&)
 //==============================================================================
 void AteshAudioProcessor::prepareToPlay (double sampleRate, int maximumExpectedSamplesPerBlock)
 {
-    dspAlgorithm.prepare (sampleRate, maximumExpectedSamplesPerBlock, getTotalNumOutputChannels());
+    dspAlgorithm.prepare (sampleRate, static_cast<unsigned int> (maximumExpectedSamplesPerBlock), static_cast<unsigned int> (getTotalNumOutputChannels()));
 }
 
 void AteshAudioProcessor::releaseResources()
